refactor: merge miem config-reading into musica::configuration
Reviewers (K20shores, boulderdaze) pointed out that src/miem/'s
ReadEmissionsConfiguration(FromString) duplicated the exact same
parse-or-throw boilerplate already in src/configuration/, just to hand
the result to a different converter. K20shores confirmed the direction:
configuration parses once into a Mechanism, then dispatches it to each
package's own converter.

Extracts that shared step into ReadMechanism/ReadMechanismFromString
(new, dependency-free of both micm and miem), and moves miem's Emissions
struct + ReadEmissionsConfiguration(FromString)/ConvertEmissions out of
src/miem/ into src/configuration/, both now built on the shared reader.
ReadConfiguration/ReadConfigurationFromString (micm side) become thin
wrappers over the same shared step; ConvertMechanism and its converters
are unchanged, so existing callers (C API, Python, JS/WASM bindings)
need no changes.

src/configuration/ was previously only built when MUSICA_ENABLE_MICM
was on; since it now also hosts emissions parsing, add_subdirectory(configuration)
moved out of that guard, with parse.cpp/emissions.cpp individually gated
by MUSICA_ENABLE_MICM/MUSICA_ENABLE_MIEM.

Verified clean builds + full ctest pass in all three configurations:
MICM+MIEM, MICM-only, MIEM-only.
