+++ src/configuration/emissions.cpp
@@ -1,6 +1,7 @@
 // Copyright (C) 2023-2026 University Corporation for Atmospheric Research
 // SPDX-License-Identifier: Apache-2.0
-#include <musica/miem/parse.hpp>
+#include <musica/configuration/emissions.hpp>
+#include <musica/configuration/read_mechanism.hpp>
 #include <musica/utils/error_code.hpp>
 
 #include <miem/species_map.hpp>
@@ -142,5 +143,15 @@ namespace musica
     emissions.regridding.type_ = ConvertRegriddingType(config.regridding.type);
 
     return emissions;
+  }
+
+  Emissions ReadEmissionsConfiguration(const std::string& config_path)
+  {
+    return ConvertEmissions(ReadMechanism(config_path));
+  }
+
+  Emissions ReadEmissionsConfigurationFromString(const std::string& json_or_yaml_string)
+  {
+    return ConvertEmissions(ReadMechanismFromString(json_or_yaml_string));
   }
 }  // namespace musica
