+++ src/configuration/parse.cpp
@@ -6,6 +6,7 @@
 #include <micm/process/rate_constant/lambda_rate_constant.hpp>
 #include <musica/micm/lambda_callback.hpp>
 #include <musica/configuration/parse.hpp>
+#include <musica/configuration/read_mechanism.hpp>
 #include <musica/utils/error_code.hpp>
 #include <sstream>
 
@@ -17,44 +18,12 @@ namespace musica
 {
   Chemistry ReadConfiguration(const std::string& config_path)
   {
-    Chemistry chemistry{};
-
-    auto parsed = mechanism_configuration::Parse(config_path);
-    if (!parsed)
-    {
-      std::string errors;
-      for (const auto& [code, message] : parsed.error())
-      {
-        errors += message + "\n";
-      }
-      throw musica::Exception(musica::ParseErrorCode::InvalidConfigFile, errors);
-    }
-    else
-    {
-      chemistry = ConvertMechanism(*parsed);
-    }
-
-    return chemistry;
+    return ConvertMechanism(ReadMechanism(config_path));
   }
 
   Chemistry ReadConfigurationFromString(const std::string& json_or_yaml_string)
   {
-    Chemistry chemistry{};
-
-    auto parsed = mechanism_configuration::ParseFromString(json_or_yaml_string);
-
-    if (!parsed)
-    {
-      std::string errors = "Failed to parse configuration string:\n";
-      for (const auto& [code, message] : parsed.error())
-      {
-        errors += message + "\n";
-      }
-      throw musica::Exception(musica::ParseErrorCode::ParsingFailed, errors);
-    }
-
-    chemistry = ConvertMechanism(*parsed);
-    return chemistry;
+    return ConvertMechanism(ReadMechanismFromString(json_or_yaml_string));
   }
 
   bool IsBool(const std::string& value)
