--- src/miem/CMakeLists.txt
@@ -1,6 +1,0 @@
-target_link_libraries(musica PUBLIC musica::miem)
-
-target_sources(musica PRIVATE
-  convert.cpp
-  parse.cpp
-)
